created independent new conditions workflow files for confirmation page
