+++ src/applications/disability-benefits/all-claims/components/confirmationFields/ConfirmationNewAndRatedConditions.jsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ConfirmationRatedDisabilities from './ConfirmationRatedDisabilities';
-import ConfirmationNewDisabilities from './ConfirmationNewDisabilities';
-
-// The ConfirmationNewDisabilities and ConfirmationRatedDisabilities are existing components.
-// ConfirmationNewConditions is a new component to manage the new conditions workflow.
+import ConfirmationDisCondNewDisabilities from './ConfirmationDisCondNewDisabilities';
+import ConfirmationDisCondRatedDisabilities from './ConfirmationDisCondRatedDisabilities';
 
 const ConfirmationNewAndRatedConditions = ({ formData }) => {
   return (
     <>
-      <ConfirmationRatedDisabilities formData={formData} />
-      <ConfirmationNewDisabilities formData={formData} />
+      <ConfirmationDisCondRatedDisabilities formData={formData} />
+      <ConfirmationDisCondNewDisabilities formData={formData} />
     </>
   );
 };
