+++ src/applications/disability-benefits/all-claims/components/confirmationFields/ConfirmationNewDisabilities.jsx
@@ -4,188 +4,145 @@ import { capitalizeEachWord } from '../../utils';
 import { formatDateString } from '../../content/conditions';
 import { DISABILITY_CAUSE_LABELS } from '../../constants';
 
-const getCauseLabel = cause => DISABILITY_CAUSE_LABELS[cause] || '';
+const getCauseLabel = cause => {
+  return DISABILITY_CAUSE_LABELS[cause] || '';
+};
 
 const ConfirmationNewDisabilities = ({ formData }) => {
   const { newDisabilities = [] } = formData || {};
 
   return (
     <>
-      {newDisabilities
-        .filter(dis => dis?.condition !== 'Rated Disability')
-        .map((dis, index) => {
-          // Guard is necessary for legacy 0781 conditions that do not have a cause
-          const cause = dis?.cause || 'Claimed';
-          const capitalizedConditionType =
-            cause === 'VA'
-              ? 'VA'
-              : cause.charAt(0).toUpperCase() + cause.slice(1).toLowerCase();
-          const causeLabel = getCauseLabel(cause);
-          const conditionDate = formatDateString(dis?.conditionDate);
-          const condition = dis?.condition?.trim();
-          const key = condition
-            ? `${cause}-${condition}`
-            : `${cause}-unknown-${index}`;
+      {newDisabilities.map(dis => {
+        // guard is necessary for legacy 0781 conditions that do not have a cause
+        const cause = dis?.cause || 'Claimed';
+        const capitalizedConditionType =
+          cause === 'VA'
+            ? 'VA'
+            : cause.charAt(0).toUpperCase() + cause.slice(1).toLowerCase();
+        const causeLabel = getCauseLabel(cause);
+        const conditionDate = formatDateString(dis?.conditionDate);
 
-          // --- Backward-compatible field accessors ---
-          // SECONDARY (new shape OR legacy nested shape)
-          const causedByDisability =
-            dis?.causedByDisability ??
-            dis?.['view:secondaryFollowUp']?.causedByDisability;
+        return (
+          <li key={dis.condition}>
+            <h4>{capitalizeEachWord(dis.condition)}</h4>
 
-          const causedByDisabilityDescription =
-            dis?.causedByDisabilityDescription ??
-            dis?.['view:secondaryFollowUp']?.causedByDisabilityDescription;
+            {capitalizedConditionType && (
+              <>
+                <div className="vads-u-color--gray">Type of condition</div>
+                <div className="vads-u-margin-bottom--2">
+                  {capitalizedConditionType} condition
+                </div>
+              </>
+            )}
 
-          // WORSENED (new shape OR legacy nested shape)
-          const worsenedDescription =
-            dis?.worsenedDescription ??
-            dis?.['view:worsenedFollowUp']?.worsenedDescription;
+            {causeLabel && (
+              <>
+                <div className="vads-u-color--gray">Cause</div>
+                <div className="vads-u-margin-bottom--2">{causeLabel}</div>
+              </>
+            )}
 
-          const worsenedEffects =
-            dis?.worsenedEffects ??
-            dis?.['view:worsenedFollowUp']?.worsenedEffects;
+            {dis.primaryDescription && (
+              <>
+                <div className="vads-u-color--gray">Description</div>
+                <div className="vads-u-margin-bottom--2">
+                  {dis.primaryDescription}
+                </div>
+              </>
+            )}
 
-          // VA (new shape OR legacy nested shape)
-          const vaMistreatmentDescription =
-            dis?.vaMistreatmentDescription ??
-            dis?.['view:vaFollowUp']?.vaMistreatmentDescription;
+            {conditionDate && (
+              <>
+                <div className="vads-u-color--gray">Date</div>
+                <div className="vads-u-margin-bottom--2">{conditionDate}</div>
+              </>
+            )}
 
-          const vaMistreatmentLocation =
-            dis?.vaMistreatmentLocation ??
-            dis?.['view:vaFollowUp']?.vaMistreatmentLocation;
-
-          const vaMistreatmentDate =
-            dis?.vaMistreatmentDate ??
-            dis?.['view:vaFollowUp']?.vaMistreatmentDate;
-
-          return (
-            <li key={key}>
-              <h4>{capitalizeEachWord(condition)}</h4>
-
-              {capitalizedConditionType && (
+            {dis.cause === 'SECONDARY' &&
+              dis?.['view:secondaryFollowUp']?.causedByDisability && (
                 <>
-                  <div className="vads-u-color--gray">Type of condition</div>
+                  <div className="vads-u-color--gray">Caused by</div>
                   <div className="vads-u-margin-bottom--2">
-                    {capitalizedConditionType} condition
+                    {dis['view:secondaryFollowUp']?.causedByDisability}
                   </div>
+                  {dis['view:secondaryFollowUp']
+                    ?.causedByDisabilityDescription && (
+                    <>
+                      <div className="vads-u-color--gray">Description</div>
+                      <div className="vads-u-margin-bottom--2">
+                        {
+                          dis['view:secondaryFollowUp']
+                            ?.causedByDisabilityDescription
+                        }
+                      </div>
+                    </>
+                  )}
                 </>
               )}
 
-              {causeLabel && (
+            {dis.cause === 'WORSENED' &&
+              dis?.['view:worsenedFollowUp'] && (
                 <>
-                  <div className="vads-u-color--gray">Cause</div>
-                  <div className="vads-u-margin-bottom--2">{causeLabel}</div>
+                  {dis['view:worsenedFollowUp'].worsenedDescription && (
+                    <>
+                      <div className="vads-u-color--gray">
+                        Worsened description
+                      </div>
+                      <div className="vads-u-margin-bottom--2">
+                        {dis['view:worsenedFollowUp'].worsenedDescription}
+                      </div>
+                    </>
+                  )}
+                  {dis['view:worsenedFollowUp'].worsenedEffects && (
+                    <>
+                      <div className="vads-u-color--gray">Worsened effects</div>
+                      <div className="vads-u-margin-bottom--2">
+                        {dis['view:worsenedFollowUp'].worsenedEffects}
+                      </div>
+                    </>
+                  )}
                 </>
               )}
 
-              {dis?.primaryDescription && (
+            {dis.cause === 'VA' &&
+              dis?.['view:vaFollowUp'] && (
                 <>
-                  <div className="vads-u-color--gray">Description</div>
-                  <div className="vads-u-margin-bottom--2">
-                    {dis.primaryDescription}
-                  </div>
+                  {dis['view:vaFollowUp'].vaMistreatmentDescription && (
+                    <>
+                      <div className="vads-u-color--gray">
+                        VA mistreatment description
+                      </div>
+                      <div className="vads-u-margin-bottom--2">
+                        {dis['view:vaFollowUp'].vaMistreatmentDescription}
+                      </div>
+                    </>
+                  )}
+                  {dis['view:vaFollowUp'].vaMistreatmentLocation && (
+                    <>
+                      <div className="vads-u-color--gray">
+                        VA mistreatment location
+                      </div>
+                      <div className="vads-u-margin-bottom--2">
+                        {dis['view:vaFollowUp'].vaMistreatmentLocation}
+                      </div>
+                    </>
+                  )}
+                  {dis['view:vaFollowUp'].vaMistreatmentDate && (
+                    <>
+                      <div className="vads-u-color--gray">
+                        VA mistreatment date
+                      </div>
+                      <div className="vads-u-margin-bottom--2">
+                        {dis['view:vaFollowUp'].vaMistreatmentDate}
+                      </div>
+                    </>
+                  )}
                 </>
               )}
-
-              {conditionDate && (
-                <>
-                  <div className="vads-u-color--gray">Date</div>
-                  <div className="vads-u-margin-bottom--2">{conditionDate}</div>
-                </>
-              )}
-
-              {/* SECONDARY */}
-              {cause === 'SECONDARY' &&
-                causedByDisability && (
-                  <>
-                    <div className="vads-u-color--gray">Caused by</div>
-                    <div className="vads-u-margin-bottom--2">
-                      {causedByDisability}
-                    </div>
-
-                    {causedByDisabilityDescription && (
-                      <>
-                        <div className="vads-u-color--gray">Description</div>
-                        <div className="vads-u-margin-bottom--2">
-                          {causedByDisabilityDescription}
-                        </div>
-                      </>
-                    )}
-                  </>
-                )}
-
-              {/* WORSENED */}
-              {cause === 'WORSENED' &&
-                (worsenedDescription || worsenedEffects) && (
-                  <>
-                    {worsenedDescription && (
-                      <>
-                        <div className="vads-u-color--gray">
-                          Worsened description
-                        </div>
-                        <div className="vads-u-margin-bottom--2">
-                          {worsenedDescription}
-                        </div>
-                      </>
-                    )}
-
-                    {worsenedEffects && (
-                      <>
-                        <div className="vads-u-color--gray">
-                          Worsened effects
-                        </div>
-                        <div className="vads-u-margin-bottom--2">
-                          {worsenedEffects}
-                        </div>
-                      </>
-                    )}
-                  </>
-                )}
-
-              {/* VA */}
-              {cause === 'VA' &&
-                (vaMistreatmentDescription ||
-                  vaMistreatmentLocation ||
-                  vaMistreatmentDate) && (
-                  <>
-                    {vaMistreatmentDescription && (
-                      <>
-                        <div className="vads-u-color--gray">
-                          VA mistreatment description
-                        </div>
-                        <div className="vads-u-margin-bottom--2">
-                          {vaMistreatmentDescription}
-                        </div>
-                      </>
-                    )}
-
-                    {vaMistreatmentLocation && (
-                      <>
-                        <div className="vads-u-color--gray">
-                          VA mistreatment location
-                        </div>
-                        <div className="vads-u-margin-bottom--2">
-                          {vaMistreatmentLocation}
-                        </div>
-                      </>
-                    )}
-
-                    {vaMistreatmentDate && (
-                      <>
-                        <div className="vads-u-color--gray">
-                          VA mistreatment date
-                        </div>
-                        <div className="vads-u-margin-bottom--2">
-                          {vaMistreatmentDate}
-                        </div>
-                      </>
-                    )}
-                  </>
-                )}
-            </li>
-          );
-        })}
+          </li>
+        );
+      })}
     </>
   );
 };
